Extract sync delta planners
